Teste task 1

+++ 0x08/resumo_pedido/Pedido.java
@@ -2,12 +2,12 @@ import provedores.Frete;
 
 public class Pedido {
 
-     int codigo;
+     double codigo;
      double peso;
      double total;
      Frete frete;
 
-    public Pedido(int codigo, double peso, double total) {
+    public Pedido(double codigo, double peso, double total) {
         this.codigo = codigo;
         this.peso = peso;
         this.total = total;
